Combine into a single module
