+++ src/app/dim-ui/VirtualList.tsx
@@ -1,9 +1,9 @@
-import { useVirtualizer } from '@tanstack/react-virtual';
+import { useVirtualizer, useWindowVirtualizer } from '@tanstack/react-virtual';
 import clsx from 'clsx';
-import { Key, useRef } from 'react';
+import { Key, useLayoutEffect, useRef } from 'react';
 import styles from './VirtualList.m.scss';
 
-export interface VirtualListProps {
+interface VirtualListProps {
   numElements: number;
   estimatedSize: number;
   className?: string;
@@ -24,7 +24,7 @@ export interface VirtualListProps {
  *
  * @see WindowVirtualList for a window-linked virtual scroller.
  */
-export default function VirtualList({
+export function VirtualList({
   numElements,
   estimatedSize,
   className,
@@ -75,6 +75,70 @@ export default function VirtualList({
             </div>
           ))}
         </div>
+      </div>
+    </div>
+  );
+}
+
+/**
+ * A virtual scrolling list linked to window scroll. e.g. Optimizer sets or Loadouts.
+ *
+ * @see VirtualList for an element-linked virtual scroller
+ */
+export function WindowVirtualList({
+  numElements,
+  estimatedSize,
+  className,
+  itemContainerClassName,
+  children,
+  overscan,
+  getItemKey,
+}: VirtualListProps) {
+  // Dynamic-height window-based virtual list code based on https://tanstack.com/virtual/v3/docs/examples/react/dynamic
+  const parentRef = useRef<HTMLDivElement>(null);
+  const parentOffsetRef = useRef(0);
+  useLayoutEffect(() => {
+    parentOffsetRef.current = parentRef.current?.offsetTop ?? 0;
+  }, []);
+
+  const virtualizer = useWindowVirtualizer({
+    count: numElements,
+    estimateSize: () => estimatedSize,
+    scrollMargin: parentOffsetRef.current,
+    getItemKey,
+    overscan,
+  });
+
+  if (numElements === 0) {
+    return null;
+  }
+
+  const items = virtualizer.getVirtualItems();
+
+  return (
+    <div
+      className={clsx(className, styles.contentsPlaceholder)}
+      ref={parentRef}
+      style={{
+        height: `${virtualizer.getTotalSize()}px`,
+      }}
+    >
+      <div
+        className={styles.virtualArea}
+        style={{
+          transform: `translateY(${items[0].start - virtualizer.options.scrollMargin}px)`,
+        }}
+      >
+        {items.map((virtualItem) => (
+          <div
+            key={virtualItem.key}
+            ref={virtualizer.measureElement}
+            className={itemContainerClassName}
+            data-index={virtualItem.index}
+          >
+            {children(virtualItem.index)}
+          </div>
+        ))}
       </div>
     </div>
   );
